Add iometer test
This commit adds both an iometer_test.sh script and an iometer jobfile
to the benchmark suite. The script can be run a non-privileged user,
and is just a wrapper to the iometer jobfile, allowing the user to
specify the load in a more friendly fashion (linear, lighter, light,
 moderate, heavy). The script also allows to specify the size of the
I/O to be performed.